Added cart realization
cart page with total sum and items, the page allows you to change
quantity of the item in cart or delete it. Proceed to checkout button
that will transfer to checkout page (when there will be one) made add to
cart buttons work on main page and product details page.