Chat service+kafka, multiple link apollo client
First of all, I created kafka consumer and producer for saving the chat
messages and did a seperate micro server for chat and database for
saving the messgaes, now when someone uses the chat it adds the message
to kafka queue and the consumer then saves it it database which reduces
the delay in the chat.

Because I created a chat microservice I deleted a chat app from the shop
django project and websocket from asgi application.

In reactjs because I now need to send gql requests for different ports I
started to use multiple client library(or someting like that) which
allows me to use 1 client and send requests to different servers.
So I also changed every gql request in reactjs and specified the server
where I want my requests to go (app or chat)